+++ eventstore/mongodb/native/src/test/java/org/occurrent/eventstore/mongodb/nativedriver/MongoEventStoreTest.java
@@ -42,6 +42,7 @@ import org.occurrent.domain.Name;
 import org.occurrent.domain.NameDefined;
 import org.occurrent.domain.NameWasChanged;
 import org.occurrent.eventstore.api.DuplicateCloudEventException;
+import org.occurrent.eventstore.api.SortBy;
 import org.occurrent.eventstore.api.WriteCondition;
 import org.occurrent.eventstore.api.WriteConditionNotFulfilledException;
 import org.occurrent.eventstore.api.blocking.EventStoreQueries;
@@ -78,6 +79,8 @@ import static org.junit.jupiter.api.condition.JRE.JAVA_8;
 import static org.occurrent.cloudevents.OccurrentCloudEventExtension.*;
 import static org.occurrent.condition.Condition.*;
 import static org.occurrent.domain.Composition.chain;
+import static org.occurrent.eventstore.api.SortBy.SortDirection.ASCENDING;
+import static org.occurrent.eventstore.api.SortBy.SortDirection.DESCENDING;
 import static org.occurrent.eventstore.api.WriteCondition.streamVersion;
 import static org.occurrent.eventstore.api.WriteCondition.streamVersionEq;
 import static org.occurrent.filter.Filter.*;
@@ -1351,6 +1354,77 @@ class MongoEventStoreTest {
                     Stream<CloudEvent> events = eventStore.all(SortBy.time(DESCENDING));
                     assertThat(deserialize(events)).containsExactly(nameDefined, nameWasChanged2, nameWasChanged1);
                 }
+
+                @Test
+                void sort_by_time_desc_and_natural_descending() {
+                    LocalDateTime now = LocalDateTime.now();
+                    NameDefined nameDefined = new NameDefined(UUID.randomUUID().toString(), now, "name");
+                    NameWasChanged nameWasChanged1 = new NameWasChanged(UUID.randomUUID().toString(), now, "name2");
+                    NameWasChanged nameWasChanged2 = new NameWasChanged(UUID.randomUUID().toString(), now.plusHours(1), "name3");
+
+                    // When
+                    persist("name1", nameDefined);
+                    persist("name3", nameWasChanged1);
+                    persist("name2", nameWasChanged2);
+
+                    // Then
+                    Stream<CloudEvent> events = eventStore.all(SortBy.time(DESCENDING).thenNatural(DESCENDING));
+                    assertThat(deserialize(events)).containsExactly(nameWasChanged2, nameWasChanged1, nameDefined);
+                }
+
+                @Test
+                void sort_by_time_desc_and_natural_ascending() {
+                    // Given
+                    LocalDateTime now = LocalDateTime.now();
+                    NameDefined nameDefined = new NameDefined(UUID.randomUUID().toString(), now, "name");
+                    NameWasChanged nameWasChanged1 = new NameWasChanged(UUID.randomUUID().toString(), now, "name2");
+                    NameWasChanged nameWasChanged2 = new NameWasChanged(UUID.randomUUID().toString(), now.plusHours(1), "name3");
+
+                    // When
+                    persist("name", nameDefined);
+                    persist("name", nameWasChanged1);
+                    persist("name", nameWasChanged2);
+
+                    // Then
+                    Stream<CloudEvent> events = eventStore.all(SortBy.time(DESCENDING).thenNatural(ASCENDING));
+                    assertThat(deserialize(events)).containsExactly(nameWasChanged2, nameDefined, nameWasChanged1);
+                }
+
+                @Test
+                void sort_by_time_desc_and_other_field_descending() {
+                    LocalDateTime now = LocalDateTime.now();
+                    NameDefined nameDefined = new NameDefined(UUID.randomUUID().toString(), now, "name");
+                    NameWasChanged nameWasChanged1 = new NameWasChanged(UUID.randomUUID().toString(), now, "name2");
+                    NameWasChanged nameWasChanged2 = new NameWasChanged(UUID.randomUUID().toString(), now.plusHours(1), "name3");
+
+                    // When
+                    persist("name", nameDefined);
+                    persist("name", nameWasChanged1);
+                    persist("name", nameWasChanged2);
+
+                    // Then
+                    Stream<CloudEvent> events = eventStore.all(SortBy.time(DESCENDING).then(STREAM_VERSION, DESCENDING));
+                    assertThat(deserialize(events)).containsExactly(nameWasChanged2, nameWasChanged1, nameDefined);
+                }
+
+                @Test
+                void sort_by_time_desc_and_other_field_ascending() {
+                    // Given
+                    LocalDateTime now = LocalDateTime.now();
+                    NameDefined nameDefined = new NameDefined(UUID.randomUUID().toString(), now, "name");
+                    NameWasChanged nameWasChanged1 = new NameWasChanged(UUID.randomUUID().toString(), now, "name2");
+                    NameWasChanged nameWasChanged2 = new NameWasChanged(UUID.randomUUID().toString(), now.plusHours(1), "name3");
+
+                    // When
+                    persist("name1", nameDefined);
+                    persist("name3", nameWasChanged1);
+                    persist("name2", nameWasChanged2);
+
+                    // Then
+                    Stream<CloudEvent> events = eventStore.all(SortBy.time(DESCENDING).then(STREAM_VERSION, ASCENDING));
+                    assertThat(deserialize(events)).containsExactly(nameWasChanged2, nameDefined, nameWasChanged1);
+                }
+
             }
 
             @Nested